Start docker image sooner
It will mitigate the case of Kaoto backend not yet ready when opening a
Kaoto file.
Previously, it was starting it only when the Kaoto editor was opened.
Now it is starting when it detects a Kaoto file in workspace or when all
other extensions with high priority have activated.

fixes #34

Signed-off-by: Aurélien Pupier <apupier@redhat.com>