Provide initial basic test
fixes #7

the test itself is very very minimal. It is testing only that an element
from the top bar and one from the Canvas are loading. I think it is
enough to catch main regressions and validate a first iteration of the
test structure. of course, more advanced use cases will need to be
covered.

Technical:
- use xvfb github action. It is required for an unknown reason. Several
other VS Code test extension are not using it. Note that it is required
for Linux only.
- there is a long wait condition, it is the time needed for docker image
to be downloaded.
- remove Windows and MacOS because cannot play test on them due to
Docker issue:
    - On MacOS, Docke rnto availabel at all
    - On Windows: docker: image operating system "linux" cannot be used
on this platform.