Avoid wipeout of content on restart with opened editors
The problem is that if the catalogs are not warmed up on first requests
from kaoto UI, the content is wiped out because Kaoto backend seems to
not waiting for the Catalog to be warmed up before returning an answer
and consequently it is providing an empty results as it knows nothing of
the content.

Not found an easy way to provide a non-regression test as it requires a
restart of VS Code (but maybe we can find some alternatives, reload of
workspaces/resources? the thing is that it requires the extension to be
deactivated and reactivated so that the kaoto backend shut down)

fixes #44

Signed-off-by: Aurélien Pupier <apupier@redhat.com>