Fix step extension loading
the trick is that the libraries were not visible to Step Extensions
because it is using micro-frontend patterns an it requires module
federation.

This commit is picking all dependencies and using them in federated
module with an `eager` state to true. This is not optimal. it can
potentially cause performance issue but none are visible currently (and
at least it is still better than crashing the editor when there is the
need for a step extension)

fixes #157

note: need to enclose the glob pattern for launching multiple tests

Signed-off-by: Aurélien Pupier <apupier@redhat.com>