chore: adapt UI tests to main branch
- `pf-topology__node__action-icon` is used for the route container
actions
too, not found a cleaner way for now to pick the second entry in the
list
- use replace instead of insert so that the test can work with both
released and main branch of kaoto. We do not care in this test where the
node is added
- the side panel needs to be awaited to be displayed on all OS now
(which seems more coherent)
- added an ugly sleep, not found another way to wait for panel to be
closed correctly. It seems like the close button is not ready
immediately even if displayed, unless it is it due to animation and the
click is sent to the wrong place?
- ignore extension recommendation: it avoids having a notification popup
on Windows for Windows Subsystem Linux extension to interfere with the
UI tests

Signed-off-by: Aurélien Pupier <apupier@redhat.com>