Implement Channel API to manipulate metadata and content of other files
first manually check version, working for several happy paths

Unit tests are not implemented because for now not found a way to have
them working due to errors related to module. Keeping them as skipped to
have a list of he most important use cases to test when e will have
found the solution and to remind we have no automated tests for that.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>