fix(Settings): Use default Settings from Kaoto
From Kaoto side, we have default settings in the
[settings.model.ts](https://github.com/KaotoIO/kaoto/blob/9ee3368e0e488853f61a84fa4e16bedadc801fda/packages/ui/src/models/settings/settings.model.ts#L23-L25)
file.

This commit uses that and only override what needs to be changed in
terms of the `catalogUrl` and the `nodeLabel`.

Once the VS Code setting for the new property `nodeToolbarTrigger` is merged, we can add it also to the VS Code settings object.