Workaround in UI test to click on contextual menu
With React 18, in one UI test, some contextual menu item
are hidden and cannot be clicked directly. Added a workaround by opening
a view which is causing a redraw of the contextual menu making it
completely visible and so with possibility to click on it.

Note that depending on the OS, there are different menus impacted.

The problem of not visible menu was already there before but not hit by
this UI test.

Signed-off-by: Aurélien Pupier <apupier@ibm.com>